requirements.txt
google-cloud-speech
google-cloud-texttospeech
precise-runner
vext
vext.gi
PyAudio~=0.2.11
six~=1.16.0
playsound~=1.2.2
numpy~=1.19.5
pip~=21.2.1
wheel~=0.36.2
cryptography~=3.4.7
mock~=4.0.3
gast~=0.5.0
PyYAML~=5.4.1
astor~=0.8.1
requests~=2.25.1
pytz~=2021.1
setuptools~=57.0.0
cffi~=1.14.5
Markdown~=3.3.4
Werkzeug~=2.0.1
dataclasses
zipp~=3.5.0
